unknown progress

+++ src/Exception.h
@@ -3,15 +3,17 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <ctype.h>
+#include <stdarg.h>
 typedef struct Exception Exception;
 typedef struct Exception* ExceptionPtr;
 
 
 struct Exception{               //If got more variable can add here
   char *msg;
-  int errorCode;
+  uint8_t errorCode;
 };
-Exception *createException(char *msg,int errorCode);
+Exception *createException(char *msg, uint8_t errorCode);
 
 void freeException(Exception *e);
 void dumpException(Exception *e);
